+++ Src/Library/changelog.md
@@ -3,3 +3,4 @@
 
 ### IMPROVEMENTS
 - swagger request & response example serialization
+- add extension method to easily obtain `EndpointDefinition` in a nswag op-processor #311